+++ app/src/main/java/com/diegowh/konohaproject/domain/timer/TimerService.kt
@@ -4,7 +4,7 @@ import android.app.Service
 import android.content.Intent
 import android.os.Binder
 import android.os.IBinder
-import com.diegowh.konohaproject.utils.ServiceNotifier
+import com.diegowh.konohaproject.utils.service.ServiceNotifier
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
