+++ app/src/main/java/com/diegowh/konohaproject/utils/timer/Interval.kt
@@ -1,0 +1,7 @@
+package com.diegowh.konohaproject.utils.timer
+
+data class Interval(
+    val currentRound: Int,
+    val isFocus: Boolean,
+    val nextDuration: Long
+)
